Make the read path work in O(N) again.
After the changes to `RowSeq' to support reading compacted rows, the
code kept using naive `DataPointsIterator', which does random access
into the data structure.  With variable length encoding, we can no
longer locate the n-th value in O(1), so the code needed a O(N) walk
to find each data point, leading to an quadratic loop when iterating
on all the data points in a `RowSeq'.

With this change, `RowSeq' uses its own custom iterator implementation
instead of `DataPointsIterator' in order to work in O(N) [single pass].

Change-Id: I98d00e1b30463542134e087243c6243e4c03ac3a