Make the compaction thread more robust.
A Throwable that isn't an Exception could kill the thread.  This is bad
as without this thread the TSD will slowly but surely run out of memory.
We can handle OOME by dropping the queue on the floor to free up memory.
Other Throwables will just cause the thread to die after 1s, but a new
one will be created to replace it.

Change-Id: I2bdb65338da43a0f366cdaa6b2b2e7aeabcd503e